Improves mobile styling
This still needs work but at least the navbar is no longer at the
bottom.

Also includes blog.zip which I used to export to Micro.blog - keeping it
around just in case but since I think I will be moving over there I'm
happy to keep this in as zip of all published posts to this point.